% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avpc-package.R
\docType{package}
\name{avpc-package}
\alias{avpc}
\alias{avpc-package}
\title{avpc: What the Package Does (One Line, Title Case)}
\description{
Perform average predictive comparisons.
}
\author{
\strong{Maintainer}: Akira Terui \email{hanabi0111@gmail.com} (\href{https://orcid.org/0000-0003-3774-2844}{ORCID})

}
\keyword{internal}
