% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_apcomp.R
\name{apcomp}
\alias{apcomp}
\title{Average predictive comparisons}
\usage{
apcomp(m, u, v = NULL, var_transform = NULL)
}
\arguments{
\item{m}{Model object. The function accepts model classes \code{"lm"}, \code{"rlm"}, \code{"glm"}, \code{"lmerMod"}}

\item{u}{Character string indicating input variable of interest.}

\item{v}{Character string indicating variables other than the input variable of interest. By default, the function uses all the variables except \code{u}.}

\item{var_transform}{Function transforming the scale of the response variable. If NULL (default), the function extracts an inverse link function from the model object \code{m}. Accepts either \code{"log"}, \code{"logit"}, or \code{"identity"}.}
}
\description{
Average predictive comparisons
}
